Move plotting functions back to this repo
Previously the band plot functions were added to Wannier.jl as package
extension. However, when adding more complex plotting functions, it is
unavoidable to introduce dependencies on other packages. Using package
extension is not ideal for managing such dependencies, also it is not
easy to auto generate the documentations for package extension functions
using Documenter.jl. Therefore, the plotting functions are now moved
back into this repo, and the user can just `using WannierPlots` to have
all the plotting functions available, instead of `using` explicitly all
the dependencies of this package to make them available if we merely
keep the functions as package extensions in Wannier.jl.